#1945868: upgrade azure-resourcemanager-appplatform to the latest and fix compatibility issues.

+++ azure-toolkit-libs/azure-toolkit-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/lib/springcloud/config/SpringCloudDeploymentConfig.java
@@ -50,10 +50,10 @@ public class SpringCloudDeploymentConfig {
 
     @Nullable
     @Builder.Default
-    private Integer cpu = 1;
+    private Double cpu = 1.0;
     @Nullable
     @Builder.Default
-    private Integer memoryInGB = 1;
+    private Double memoryInGB = 1.0;
     private Integer instanceCount;
     private String deploymentName;
     @Nullable
